+++ tests/conftest.py
@@ -5139,6 +5139,239 @@ def populated_profiler_store(
 
 
 @pytest.fixture
+@freeze_time("09/26/2019 13:42:41")
+def alice_columnar_table_single_batch(empty_data_context):
+    """
+    About the "Alice" User Workflow Fixture
+
+    Alice has a single table of columnar data called user_events (DataAsset) that she wants to check periodically as new
+    data is added.
+
+      - She knows what some of the columns mean, but not all - and there are MANY of them (only a subset currently shown
+        in examples and fixtures).
+
+      - She has organized other tables similarly so that for example column name suffixes indicate which are for user
+        ids (_id) and which timestamps are for versioning (_ts).
+
+    She wants to use a configurable profiler to generate a description (ExpectationSuite) about table so that she can:
+
+        1. use it to validate the user_events table periodically and set up alerts for when things change
+
+        2. have a place to add her domain knowledge of the data (that can also be validated against new data)
+
+        3. if all goes well, generalize some of the Profiler to use on her other tables
+
+    Alice configures her Profiler using the YAML configurations and data file locations captured in this fixture.
+    """
+    skip_if_python_below_minimum_version()
+
+    verbose_profiler_config_file_path: str = file_relative_path(
+        __file__,
+        os.path.join(
+            "test_fixtures",
+            "rule_based_profiler",
+            "alice_user_workflow_verbose_profiler_config.yml",
+        ),
+    )
+
+    verbose_profiler_config: str
+    with open(verbose_profiler_config_file_path) as f:
+        verbose_profiler_config = f.read()
+
+    my_rule_for_user_ids_expectation_configurations: List[ExpectationConfiguration] = [
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_values_to_be_of_type",
+                "kwargs": {
+                    "column": "user_id",
+                    "type_": "INTEGER",
+                },
+                "meta": {},
+            }
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_values_to_be_between",
+                "kwargs": {
+                    "min_value": 397433,  # From the data
+                    "max_value": 999999999999,
+                    "column": "user_id",
+                },
+                "meta": {},
+            }
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_values_to_not_be_null",
+                "kwargs": {
+                    "column": "user_id",
+                },
+                "meta": {},
+            }
+        ),
+    ]
+
+    event_ts_column_data: Dict[str, str] = {
+        "column_name": "event_ts",
+        "observed_max_time_str": "2004-10-19 11:05:20",
+        "observed_strftime_format": "%Y-%m-%d %H:%M:%S",
+    }
+
+    my_rule_for_timestamps_column_data: List[Dict[str, str]] = [
+        event_ts_column_data,
+        {
+            "column_name": "server_ts",
+            "observed_max_time_str": "2004-10-19 11:05:20",
+        },
+        {
+            "column_name": "device_ts",
+            "observed_max_time_str": "2004-10-19 11:05:22",
+        },
+    ]
+    my_rule_for_timestamps_expectation_configurations: List[
+        ExpectationConfiguration
+    ] = []
+    column_data: Dict[str, str]
+    for column_data in my_rule_for_timestamps_column_data:
+        my_rule_for_timestamps_expectation_configurations.extend(
+            [
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_values_to_be_of_type",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                            "type_": "TIMESTAMP",
+                        },
+                        "meta": {},
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_values_to_be_increasing",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                        },
+                        "meta": {},
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_values_to_be_dateutil_parseable",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                        },
+                        "meta": {},
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_min_to_be_between",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                            "min_value": "2004-10-19T10:23:54",  # From variables
+                            "max_value": "2004-10-19T10:23:54",  # From variables
+                        },
+                        "meta": {
+                            "notes": {
+                                "format": "markdown",
+                                "content": [
+                                    "### This expectation confirms no events occur before tracking started **2004-10-19 10:23:54**"
+                                ],
+                            }
+                        },
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_max_to_be_between",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                            "min_value": "2004-10-19T10:23:54",  # From variables
+                            "max_value": event_ts_column_data[
+                                "observed_max_time_str"
+                            ],  # Pin to event_ts column
+                        },
+                        "meta": {
+                            "notes": {
+                                "format": "markdown",
+                                "content": [
+                                    "### This expectation confirms that the event_ts contains the latest timestamp of all domains"
+                                ],
+                            }
+                        },
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_values_to_match_strftime_format",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                            "strftime_format": {
+                                "value": event_ts_column_data[
+                                    "observed_strftime_format"
+                                ],  # Pin to event_ts column
+                                "details": {"success_ratio": 1.0},
+                            },
+                        },
+                        "meta": {
+                            "notes": {
+                                "format": "markdown",
+                                "content": [
+                                    "### This expectation confirms that fields ending in _ts are of the format detected by parameter builder SimpleDateFormatStringParameterBuilder"
+                                ],
+                            }
+                        },
+                    }
+                ),
+            ]
+        )
+
+    expectation_configurations: List[ExpectationConfiguration] = []
+
+    expectation_configurations.extend(my_rule_for_user_ids_expectation_configurations)
+    expectation_configurations.extend(my_rule_for_timestamps_expectation_configurations)
+
+    expectation_suite_name: str = "alice_columnar_table_single_batch"
+    expected_expectation_suite: ExpectationSuite = ExpectationSuite(
+        expectation_suite_name=expectation_suite_name, data_context=empty_data_context
+    )
+    expectation_configuration: ExpectationConfiguration
+    for expectation_configuration in expectation_configurations:
+        # NOTE Will 20211208 add_expectation() method, although being called by an ExpectationSuite instance, is being
+        # called within a fixture, and we will prevent it from sending a usage_event by calling the private method
+        # _add_expectation().
+        expected_expectation_suite._add_expectation(
+            expectation_configuration=expectation_configuration, send_usage_event=False
+        )
+
+    # NOTE that this expectation suite should fail when validated on the data in "sample_data_relative_path"
+    # because the device_ts is ahead of the event_ts for the latest event
+    sample_data_relative_path: str = "alice_columnar_table_single_batch_data.csv"
+
+    profiler_config: dict = yaml.load(verbose_profiler_config)
+
+    # Roundtrip through schema validation to remove any illegal fields add/or restore any missing fields.
+    deserialized_config: dict = ruleBasedProfilerConfigSchema.load(profiler_config)
+    serialized_config: dict = ruleBasedProfilerConfigSchema.dump(deserialized_config)
+
+    # `class_name`/`module_name` are generally consumed through `instantiate_class_from_config`
+    # so we need to manually remove those values if we wish to use the **kwargs instantiation pattern
+    serialized_config.pop("class_name")
+    serialized_config.pop("module_name")
+    expected_expectation_suite.add_citation(
+        comment="Suite created by Rule-Based Profiler with the configuration included.",
+        profiler_config=serialized_config,
+    )
+
+    return {
+        "profiler_config": verbose_profiler_config,
+        "expected_expectation_suite_name": expectation_suite_name,
+        "expected_expectation_suite": expected_expectation_suite,
+        "sample_data_relative_path": sample_data_relative_path,
+    }
+
+
+@pytest.fixture
 def alice_columnar_table_single_batch_context(
     monkeypatch,
     empty_data_context,
@@ -6462,138 +6695,120 @@ def alice_columnar_table_single_batch(empty_data_context):
     }
 
 
-# TODO: AJB 20210525 This fixture is not yet used but may be helpful to generate batches for unit tests of multibatch
-#  workflows.  It should probably be extended to add different column types / data.
 @pytest.fixture
-def multibatch_generic_csv_generator():
+def quentin_columnar_table_multi_batch():
     """
-    Construct a series of csv files with many data types for use in multibatch testing
+    About the "Quentin" User Workflow Fixture
+    Quentin has multiple tables of columnar data called user_events (DataAsset) that he wants to check periodically as
+    new data is added.
+      - He knows what some of the columns are of the accounting/financial/account type, but he is currently interested
+        in the range of quantiles of columns capturing financial quantities (column names ending on "_amount" suffix).
+    He wants to use a configurable profiler to generate a description (ExpectationSuite) about tables so that he can:
+        1. monitor the range of quantiles of columns capturing financial quantities in the tables
+        2. have a place to add his domain knowledge of the data (that can also be validated against new data)
+        3. if all goes well, generalize some of the Profiler to use on his other tables
+    Quentin uses a custom implementation of the "bootstrap" non-parametric (i.e, data-driven) statistical estimator.
+    Quentin configures his Profiler using the YAML configurations and data file locations captured in this fixture.
     """
     skip_if_python_below_minimum_version()
 
-    def _multibatch_generic_csv_generator(
-        data_path: str,
-        start_date: Optional[datetime.datetime] = None,
-        num_event_batches: Optional[int] = 20,
-        num_events_per_batch: Optional[int] = 5,
-    ) -> List[str]:
+    verbose_profiler_config_file_path: str = file_relative_path(
+        __file__,
+        os.path.join(
+            "test_fixtures",
+            "rule_based_profiler",
+            "quentin_user_workflow_verbose_profiler_config.yml",
+        ),
+    )
 
-        if start_date is None:
-            start_date = datetime.datetime(2000, 1, 1)
+    verbose_profiler_config: str
+    with open(verbose_profiler_config_file_path) as f:
+        verbose_profiler_config = f.read()
 
-        file_list = []
-        category_strings = {
-            0: "category0",
-            1: "category1",
-            2: "category2",
-            3: "category3",
-            4: "category4",
-            5: "category5",
-            6: "category6",
-        }
-        for batch_num in range(num_event_batches):
-            # generate a dataframe with multiple column types
-            batch_start_date = start_date + datetime.timedelta(
-                days=(batch_num * num_events_per_batch)
-            )
-            # TODO: AJB 20210416 Add more column types
-            df = pd.DataFrame(
-                {
-                    "event_date": [
-                        (batch_start_date + datetime.timedelta(days=i)).strftime(
-                            "%Y-%m-%d"
-                        )
-                        for i in range(num_events_per_batch)
-                    ],
-                    "batch_num": [batch_num + 1 for _ in range(num_events_per_batch)],
-                    "string_cardinality_3": [
-                        category_strings[i % 3] for i in range(num_events_per_batch)
-                    ],
-                }
-            )
-            filename = f"csv_batch_{batch_num + 1:03}_of_{num_event_batches:03}.csv"
-            file_list.append(filename)
-            # noinspection PyTypeChecker
-            df.to_csv(
-                os.path.join(data_path, filename),
-                index_label="intra_batch_index",
-            )
+    expectation_suite_name_bootstrap_sampling_method: str = (
+        "quentin_columnar_table_multi_batch"
+    )
 
-        return file_list
-
-    return _multibatch_generic_csv_generator
+    return {
+        "profiler_config": verbose_profiler_config,
+        "test_configuration": {
+            "expectation_suite_name": expectation_suite_name_bootstrap_sampling_method,
+            "expect_column_quantile_values_to_be_between_quantile_ranges_by_column": {
+                "tolls_amount": [[0.0, 0.0], [0.0, 0.0], [0.0, 0.0]],
+                "fare_amount": [
+                    [5.842754275, 6.5],
+                    [8.675167517, 9.661311131],
+                    [13.344354435, 15.815389039],
+                ],
+                "tip_amount": [
+                    [0.0, 0.0],
+                    [0.81269502, 1.97259736],
+                    [2.346049055, 2.993680968],
+                ],
+                "total_amount": [
+                    [8.2740033, 11.422183043],
+                    [11.358555106, 14.959993149],
+                    [16.746263451, 21.327684643],
+                ],
+            },
+        },
+    }
 
 
 @pytest.fixture
-def multibatch_generic_csv_generator_context(monkeypatch, empty_data_context):
+def quentin_columnar_table_multi_batch_data_context(
+    tmp_path_factory,
+    monkeypatch,
+) -> DataContext:
+    """
+    This fixture generates three years' worth (36 months; i.e., 36 batches) of taxi trip data with the number of rows
+    of each batch being equal to the original number per log file (10,000 rows).
+    """
     skip_if_python_below_minimum_version()
 
-    context: DataContext = empty_data_context
-    monkeypatch.chdir(context.root_directory)
-    data_relative_path = "../data"
-    data_path = os.path.join(context.root_directory, data_relative_path)
-    os.makedirs(data_path, exist_ok=True)
+    # Re-enable GE_USAGE_STATS
+    monkeypatch.delenv("GE_USAGE_STATS")
 
-    data_connector_base_directory = "./"
-    monkeypatch.setenv("base_directory", data_connector_base_directory)
-    monkeypatch.setenv("data_fixtures_root", data_relative_path)
+    project_path: str = str(tmp_path_factory.mktemp("taxi_data_context"))
+    context_path: str = os.path.join(project_path, "great_expectations")
+    os.makedirs(os.path.join(context_path, "expectations"), exist_ok=True)
+    data_path: str = os.path.join(context_path, "..", "data")
+    os.makedirs(os.path.join(data_path), exist_ok=True)
+    shutil.copy(
+        file_relative_path(
+            __file__,
+            os.path.join(
+                "integration",
+                "fixtures",
+                "yellow_tripdata_pandas_fixture",
+                "great_expectations",
+                "great_expectations.yml",
+            ),
+        ),
+        str(os.path.join(context_path, "great_expectations.yml")),
+    )
+    base_directory: str = file_relative_path(
+        __file__,
+        os.path.join(
+            "test_sets",
+            "taxi_yellow_tripdata_samples",
+        ),
+    )
+    file_name_list: List[str] = get_filesystem_one_level_directory_glob_path_list(
+        base_directory_path=base_directory, glob_directive="*.csv"
+    )
+    file_name_list = sorted(file_name_list)
 
-    datasource_name = "generic_csv_generator"
-    data_connector_name = "daily_data_connector"
-    asset_name = "daily_data_asset"
-    datasource_config = rf"""
-class_name: Datasource
-module_name: great_expectations.datasource
-execution_engine:
-  module_name: great_expectations.execution_engine
-  class_name: PandasExecutionEngine
-data_connectors:
-  {data_connector_name}:
-    class_name: ConfiguredAssetFilesystemDataConnector
-    assets:
-      {asset_name}:
-        module_name: great_expectations.datasource.data_connector.asset
-        group_names:
-          - batch_num
-          - total_batches
-        pattern: csv_batch_(\d.+)_of_(\d.+)\.csv
-        reader_options:
-          delimiter: ","
-        class_name: Asset
-        base_directory: $data_fixtures_root
-        glob_directive: "*.csv"
-    base_directory: $base_directory
-    module_name: great_expectations.datasource.data_connector
-        """
+    file_name: str
+    csv_source_path: str
+    for file_name in file_name_list:
+        csv_source_path = os.path.join(base_directory, file_name)
+        shutil.copy(
+            csv_source_path,
+            os.path.join(context_path, "..", "data", file_name),
+        )
 
-    context.add_datasource(name=datasource_name, **yaml.load(datasource_config))
+    context: DataContext = DataContext(context_root_dir=context_path)
+    assert context.root_directory == context_path
 
-    assert context.list_datasources() == [
-        {
-            "class_name": "Datasource",
-            "data_connectors": {
-                data_connector_name: {
-                    "assets": {
-                        asset_name: {
-                            "base_directory": data_relative_path,
-                            "class_name": "Asset",
-                            "glob_directive": "*.csv",
-                            "group_names": ["batch_num", "total_batches"],
-                            "module_name": "great_expectations.datasource.data_connector.asset",
-                            "pattern": "csv_batch_(\\d.+)_of_(\\d.+)\\.csv",
-                        }
-                    },
-                    "base_directory": data_connector_base_directory,
-                    "class_name": "ConfiguredAssetFilesystemDataConnector",
-                    "module_name": "great_expectations.datasource.data_connector",
-                }
-            },
-            "execution_engine": {
-                "class_name": "PandasExecutionEngine",
-                "module_name": "great_expectations.execution_engine",
-            },
-            "module_name": "great_expectations.datasource",
-            "name": "generic_csv_generator",
-        }
-    ]
     return context
